Use is_root to test for root node.
Add an additional member, named "root", to the list representing the
root node when constructing the tree object using make_tree.  Other code
can then use the is_root function to test for the root node.  The
is_root function itself simply checks a given node's list representation
for the existence of a member named "root".  A node is the root node if
none of its parents belong to the tree.