When adding children ignore parent if not in tree.
When building trees from data frames with make_tree, we want to conserve
as much information as possible.  Therefore it is okay if a parent is
mentioned in the "parent" member of a child node even though the parent
itself does not belong to the tree that is being built.  That works as
long as parents that are not in the tree are skipped when children are
added to their parent's node.  Not skipping them would create incomplete
nodes, accessible under the given parent's id, with nothing but a
"children" member in their list representations.