Include original data frame in tree objects.
Hardcode column names "id" and "parents" into "make_tree" function, so
that "make_tree" always knows where to find the node and parent ids.
The alternative would be to not just include the original data frame but
also the names of the id and parents columns with the tree object and to
pass them explicitly in every call to "make_tree".  As a consequence
every data frame that is to be converted into a tree object must have
its node and parent ids in columns called "id" and "parents",
respectively.